code for palindrome number

+++ Leetcode-Easy/9-PalindromeNumber/palindromesolution.cpp
@@ -8,7 +8,7 @@ public:
             return false;
         }
         int temp=x;
-        int rev=0;
+        long int rev=0;
         while(temp!=0){
             int rem=temp%10;
             temp=temp/10;
